// src/service/file/schemas/fileObjectResult.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * File API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */
import type { FileObjectResultSource } from './fileObjectResultSource';
import type { FileObjectResultState } from './fileObjectResultState';

/**
 * 실제 썸네일 파일 정보
 */
export interface FileObjectResult {
  /** 파일의 id */
  id: number;
  /** 자료의 기본 종류 구분. FILE, THUMB, PROFILE */
  source: FileObjectResultSource;
  /** base url */
  host: string;
  /** s3 버켓 정보 */
  bucket: string;
  /** s3내에서의 키 값 */
  key: string;
  /** 미디어 타입 */
  mediaType: string;
  /** 업로드한 원본 파일의 이름 */
  originalFileName: string;
  /** 연결된 드라이브 항목 수 */
  driveItemCount: number;
  /** 배포(공유)한 항목 수 */
  publishedItemCount: number;
  /** 파일의 사이즈 */
  size: number;
  /** 파일의 업로드 상태, 0:업로드 전, 1: 정상(업로드완료. 다운/공유가능), 2: 삭제대기 */
  state: FileObjectResultState;
  /** 생성 / 업로드 한 계정의 id */
  ownerAccountId: number;
  /** 생성 / 업로드 한 ip */
  createdIp: string;
  /** 생성 / 업로드 한 날짜 */
  createdAt: string;
  /** presigned url. 업로드 시 url 정보 */
  presignedUrl?: string;
  /** stt로 전체 텍스트화 한 내용 */
  sttFullString?: string;
  /** 입려된 stt 전체 내용을를 요약한 내용 */
  sttSummary?: string;
}