// src/service/file/schemas/storyBoardContentForAdd.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * File API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */
import type { StoryBoardContentForAddContentType } from './storyBoardContentForAddContentType';
import type { StoryBoardContentPhotoForAdd } from './storyBoardContentPhotoForAdd';

/**
 * 스토리보드에 들어갈 내용들
 */
export interface StoryBoardContentForAdd {
  /**
   * 내용의 순서. 0부터 오름차순
   * @minimum 0
   */
  contentOrder: number;
  /** 내용의 타입, enum이라 0부터 넣거나 텍스트로 입력 */
  contentType: StoryBoardContentForAddContentType;
  /**
   * 제목
   * @maxLength 45
   */
  title?: string;
  /**
   * 부제목, 혹은 내용
   * @maxLength 1000
   */
  contents?: string;
  attachedPhotos?: StoryBoardContentPhotoForAdd[];
}