// src/service/file/schemas/storyBoardContentForAddContentType.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * File API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */

/**
 * 내용의 타입, enum이라 0부터 넣거나 텍스트로 입력
 */
export type StoryBoardContentForAddContentType =
  (typeof StoryBoardContentForAddContentType)[keyof typeof StoryBoardContentForAddContentType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const StoryBoardContentForAddContentType = {
  TYPE_A: 'TYPE_A',
  TYPE_B: 'TYPE_B',
  TYPE_C: 'TYPE_C',
  TYPE_D: 'TYPE_D',
  TYPE_E: 'TYPE_E',
  TYPE_F: 'TYPE_F',
} as const;