// src/service/file/schemas/storyBoardContentForUpdate.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * File API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */
import type { StoryBoardContentForUpdateContentType } from './storyBoardContentForUpdateContentType';
import type { StoryBoardContentPhotoForUpdate } from './storyBoardContentPhotoForUpdate';

/**
 * 스토리보드에 들어갈 내용들
 */
export interface StoryBoardContentForUpdate {
  /**
   * 컨텐츠의 id 기존의 id가 있다면 전달.(새로 추가한 것들은 없을테니)
   * @nullable
   */
  storyBoardContentId?: number | null;
  /**
   * 내용의 순서. 0부터 오름차순
   * @minimum 0
   */
  contentOrder: number;
  /** 내용의 타입, enum이라 0부터 넣거나 텍스트로 입력 */
  contentType: StoryBoardContentForUpdateContentType;
  /**
   * 제목
   * @maxLength 45
   */
  title?: string;
  /**
   * 부제목, 혹은 내용
   * @maxLength 1000
   */
  contents?: string;
  attachedPhotos?: StoryBoardContentPhotoForUpdate[];
}