// src/service/file/schemas/storyBoardContentPhotoForAdd.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * File API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */

export interface StoryBoardContentPhotoForAdd {
  /**
   * 포함되는 자료의 id(driveItem.id)나 키(driveItemKey)
   * @maxLength 36
   */
  driveItemIdOrKey: string;
  /** 사진의 순서. 기존에 중간에 빈곳이 있다면 null로 보냈어야했는데, 순서로 넣어주시면 될 것 같습니다. */
  order: number;
  /** 사진 위치조정을 위한 x 값 */
  positionX: number;
  /** 사진 위치조정을 위한 y 값 */
  positionY: number;
  /** 사진 위치조정을 위한 가로 값 */
  width: number;
  /** 사진 위치조정을 위한 세로 값 */
  height: number;
}