// src/service/file/schemas/storyBoardContentPhotoResult.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * File API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */

/**
 * 포함된 사진의 리스트
 */
export interface StoryBoardContentPhotoResult {
  /** 스토리보드 사진의 아이디 */
  id: number;
  /** 속한 스토리보드의 모듈? 컨텐츠 아이디 */
  contentId: number;
  /** 사진의 드라이브아이템 키 값 */
  driveItemKey: string;
  /** 사진의 URL */
  photoUrl: string;
  /** 사진의 순서. 입력시 들어온 순. 0부터 오름차순 */
  order: number;
  /** 사진의 x좌표 */
  positionX: number;
  /** 사진의 y좌표 */
  positionY: number;
  /** 사진의 가로 크기 */
  width: number;
  /** 사진의 세로 크기 */
  height: number;
}